bpflbr: Output func typed args and typed retval
It's really useful to check args and return value details while tracing
bpf progs and kernel functions.

As for args, it's really cool to output arg's type, name and value.

As for return value, it's better to output type and value than only output
untyped value.

In order to achieve it, bpflbr uses bpf_get_func_arg() helper, which
limits tracee has no more than 6 args.

If the arg is a string, aka const char *, bpflbr is able to read the str
as the arg's value.

If the arg is a enum/int pointer, bpflbr is able to read the pointed
data as the arg's second value.

How does bpf code know the arg info?

Yeah, the info is prepared by Go from the corresponding func's BTF info.

And when output args and return value, it's to output them based on
their corresponding BTF info.

No black magic behind BTF.

Signed-off-by: Leon Hwang <leon.hwang@linux.dev>